tests: split up into separate files
