+++ crates/holochain_integrity_types/src/action.rs
@@ -483,7 +483,7 @@ pub struct AgentValidationPkg {
     pub membrane_proof: Option<MembraneProof>,
 }
 
-/// A action which declares that all zome init functions have successfully
+/// An action which declares that all zome init functions have successfully
 /// completed, and the chain is ready for commits. Contains no explicit data.
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, SerializedBytes, Hash)]
 #[cfg_attr(
@@ -570,7 +570,7 @@ pub struct CloseChain {
     pub new_dna_hash: DnaHash,
 }
 
-/// A action which "speaks" Entry content into being. The same content can be
+/// An action which "speaks" Entry content into being. The same content can be
 /// referenced by multiple such actions.
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, SerializedBytes, Hash)]
 #[cfg_attr(
@@ -589,7 +589,7 @@ pub struct Create<W = EntryRateWeight> {
     pub weight: W,
 }
 
-/// A action which specifies that some new Entry content is intended to be an
+/// An action which specifies that some new Entry content is intended to be an
 /// update to some old Entry.
 ///
 /// This action semantically updates an entry to a new entry.
